Change project and issue
I greatly overestimated my free time at the beginning of the year.
Working on a compiler is really hard. While I enjoy doing it, I don't
have enough time to to this before the deadline.

As such, I think the best thing to do could be to change the issue I'll
work on in order to guarantee that I could finish it properly.